Create geopackages containing an OSMnx network
geopackages (gpkg) include information about the flow recorded by a detector, as well as whether a node is actually a detector via a size attribute. If size = 0, then the node is a non-detector node.
Also added flow values to the respective detectors in coords_matched.csv to allow coloring of matched detector nodes.